requirements.txt
flask==3.0.0
python-dotenv==1.0.0
langchain==0.1.14
langchain-community==0.0.24
langchain-openai==0.1.3
openai==1.3.0
faiss-cpu==1.7.4
pypdf==4.0.0
unstructured==0.12.0
pillow==10.0.0
pptx==0.6.21
